@achille-roussel 2017-06-14
Follow up of #9