@rbranson 2017-06-16
@achille-roussel not sure how I feel about getting rid of reading from /dev/urandom, given that this is the way that the UUID generator code I found during my research approached getting their random bits. Even though I see where you're going with just seeding the generator with /dev/urandom, I don't think I'm qualified to know that this is going to degrade the randomness of this system. This is measurable though using a Diehard test to analyze the quality of the output. The lack of collisions is one of the key properties we need to preserve in KSUID, so this is an area to approach with extreme caution.

In terms of performance, did you consider adding a buffer that reduced the overhead of /dev/urandom? Presumably most of the cost there is in the actual overhead of making the I/O call. If this can be read in, say, 16KB chunks (or bigger?), and then parceled out into pieces using a simple slice read and bumping an offset it seems like the fast path can be cheap.