@achille-roussel 2017-06-16
I had not considered using a buffer to amortize the cost of syscalls, this seems like worth the experiment (and may even give better overall performance).
I'll try it out and let you know what the results are.

About the quality of the random sequence, I'll add a test to verify it.