@achille-roussel 2017-06-20
I've made a change to get back the 2 bits of randomness that were lost by using the default random number generator. 

From everything I've read using the Ziggurat PRNG will give us the same uniqueness guarantees than reading from /dev/urandom. The only critical point is on generating the seed, which would cause duplicates if two hosts were to generate the same seed within the same second. The seed is 8 bytes so the risk is still pretty low.