@achille-roussel 2017-06-23
Following @ejcx's wise advise, cycled back to keep /dev/urandom as the default source and added `ksuid.FastRander` as an alternative to optimize KSUID generation when needed.